refactor(package): 패키지 분리
- model : Lotto 관련
- view : InputView,OutputView
- controller: Controller
- error: LottoErrorCode
- util: Parser
