+++ src/main/java/lotto/Application.java
@@ -1,5 +1,13 @@
 package lotto;
 
+import lotto.controller.Controller;
+import lotto.model.LottoMachine;
+import lotto.model.RandomLottoMachine;
+import lotto.view.InputView;
+import lotto.view.InputViewImpl;
+import lotto.view.OutputView;
+import lotto.view.OutputViewImpl;
+
 public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputViewImpl();
