+++ src/main/java/lotto/view/OutputView.java
@@ -1,6 +1,8 @@
-package lotto;
+package lotto.view;
 
 import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.WinningResult;
 
 public interface OutputView {
     void printTickets(List<Lotto> lottoTickets);
