+++ src/main/java/lotto/view/ViewFormatter.java
@@ -1,6 +1,8 @@
-package lotto;
+package lotto.view;
 
 import java.util.List;
+import lotto.model.Lotto;
+import lotto.model.WinningResult;
 
 public class ViewFormatter {
     public static final String ERROR_PREFIX = "[ERROR] ";
